backend/requirements.txt
flask
flask-cors
pymupdf
groq
python-dotenv
gunicorn